chore: move explorer node to its own file
